refactor for readability and fill descriptions
